Reimplement tree procesess
They construct a binary tree (or multiple) and
return references to the respective roots.

Observables A and C are stored in the val fields
of the tree. Calculate them with A! and C!, or call
treevalues! which fills the tree and
returns the tuple (A,C) of vectors A,C.

Adapt run.jl to new API.